feat(EPIC-001): Phase 0 — decimals unification and u128 token balances
Remove CBE constants from lib-types (CBE is a DAO, not a system token).
Unify SOV to 18 decimals everywhere. Widen TokenContract balances,
supply, and all amount parameters from u64 to u128.

Changes:
- lib-types/tokenomics.rs: remove CBE_DECIMALS, CBE_MAX_SUPPLY,
  CBE_TOTAL_SUPPLY_TOKENS (moved to cbe_token.rs as self-contained)
- constants.rs: SOV_TOKEN_DECIMALS 8 → 18, SOV_TOKEN_MAX_SUPPLY u64 → u128
- core.rs: TokenContract balances/supply/amounts all u128
- canonical.rs: bonding curve MAX_SUPPLY defined locally (was CBE_MAX_SUPPLY)
- 37 files updated to propagate u128 through callers

1786 lib-blockchain tests pass, 0 failures.

Refs: #2093, #2094

+++ lib-blockchain/src/blockchain/test_utils.rs
@@ -135,14 +135,14 @@ impl Blockchain {
             .get_mut(&sov_id)
             .ok_or_else(|| anyhow::anyhow!("SOV token contract not found"))?;
         token
-            .credit_balance(&pk, amount)
+            .credit_balance(&pk, amount as u128)
             .map_err(|e| anyhow::anyhow!("Treasury credit failed: {}", e))?;
         Ok(())
     }
 
     /// Query the raw SOV balance for an arbitrary 64-char hex wallet ID.
     /// For unit tests only.
-    pub fn get_wallet_sov_for_test(&self, wallet_id_hex: &str) -> Result<u64> {
+    pub fn get_wallet_sov_for_test(&self, wallet_id_hex: &str) -> Result<u128> {
         let id_bytes: [u8; 32] = hex::decode(wallet_id_hex)
             .map_err(|e| anyhow::anyhow!("Bad wallet hex: {}", e))?
             .try_into()
@@ -200,7 +200,7 @@ impl Blockchain {
             .get_mut(&sov_id)
             .ok_or_else(|| anyhow::anyhow!("SOV token contract not found"))?;
         token
-            .credit_balance(&pk, amount)
+            .credit_balance(&pk, amount as u128)
             .map_err(|e| anyhow::anyhow!("Identity SOV credit failed: {}", e))?;
         Ok(())
     }
